feat: 주문 관련 서비스 로직에 상품옵션조합 반영하여 수정

+++ src/main/java/turtleMart/order/dto/response/TotalOrderedQuantityResponse.java
@@ -2,9 +2,9 @@ package turtleMart.order.dto.response;
 
 public record TotalOrderedQuantityResponse(
         Long productId,
-        Integer totalOrdered
+        Long totalOrdered
 ) {
-    public static TotalOrderedQuantityResponse from(Long productId, Integer totalOrdered){
+    public static TotalOrderedQuantityResponse from(Long productId, Long totalOrdered){
         return new TotalOrderedQuantityResponse(productId, totalOrdered);
     }
 }
